Implementação inicial do grud de campeonatos

+++ src/main/resources/db/migration/V2__create-table-jogardores.sql
@@ -1,5 +1,5 @@
 CREATE TABLE jogadores (
-	id int AUTO_INCREMENT NOT NULL UNIQUE,
+	id BIGINT AUTO_INCREMENT NOT NULL UNIQUE,
 	cpf varchar(14) NOT NULL,
 	rg varchar(12),
 	nome varchar(130) NOT NULL,
@@ -9,7 +9,7 @@ CREATE TABLE jogadores (
 	pe_predominante varchar(10) NOT NULL DEFAULT 'Direito',
 	descricao varchar(255),
 	celular varchar(14) not null,
-	endereco_id int,
+	endereco_id BIGINT,
 	data_criacao DATETIME NOT NULL,
 	PRIMARY KEY (id)
 );
